% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{new_consonant}
\alias{new_consonant}
\title{Constructor for the consonant (sub)class.}
\usage{
new_consonant(x)
}
\arguments{
\item{x}{A named list containing IPA data.}
}
\value{
A consonant (sub)class object.
}
\description{
Constructor for the consonant (sub)class.
}
