% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{new_phoneme}
\alias{new_phoneme}
\title{Constructor for the phoneme class.}
\usage{
new_phoneme(x, class = character())
}
\arguments{
\item{x}{A named list containing IPA data.}

\item{class}{A single-element character vector - "consonant" or "vowel".}
}
\value{
A phoneme class object.
}
\description{
Constructor for the phoneme class.
}
