% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{new_vowel}
\alias{new_vowel}
\title{Constructor for the vowel (sub)class.}
\usage{
new_vowel(x)
}
\arguments{
\item{x}{A list containing IPA data.}
}
\value{
A vowel (sub)class object
}
\description{
Constructor for the vowel (sub)class.
}
\keyword{internal}
