% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{phoneme}
\alias{phoneme}
\title{Creates a new phoneme class object.}
\usage{
phoneme(ipa)
}
\arguments{
\item{ipa}{A named list referencing the phoneme's IPA representation and articulation.}
}
\value{
A phoneme class object.
}
\description{
Creates a new phoneme class object.
}
