% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{validate_phoneme}
\alias{validate_phoneme}
\title{Validates a phoneme class object.}
\usage{
validate_phoneme(x)
}
\arguments{
\item{x}{A potentially-valid phoneme class object.}
}
\value{
A valid phoneme class object.
}
\description{
Validates a phoneme class object.
}
\keyword{internal}
