% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{vowel}
\alias{vowel}
\title{Title}
\usage{
vowel(ipa)
}
\arguments{
\item{ipa}{A named list referencing the consonant's IPA representation and
articulation as length 1 and 2 character vectors respectively.}
}
\value{
A vowel object.
}
\description{
Title
}
