% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{consonant}
\alias{consonant}
\title{Creates a new consonant object.}
\usage{
consonant(ipa)
}
\arguments{
\item{ipa}{A named list referencing the vowel's IPA representation and
articulation as length 1 and 2 character vectors respectively.}
}
\value{
A consonant object.
}
\description{
Creates a new consonant object.
}
