% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{vowel}
\alias{vowel}
\title{Create a new vowel object.}
\usage{
vowel(ipa)
}
\arguments{
\item{ipa}{A list of two character vectors. The first element provides
the IPA representation for the desired phoneme and the second provides its
articulation as a character vector of length 3.}
}
\value{
A consonant object.
}
\description{
Takes IPA data, as a list of character vectors, returning a vowel object.
A vowel object is a subclass of the more general phoneme object.
}
\examples{
ipa_data <- list("e", c("close-mid", "front", "unrounded"))
e <- vowel(ipa_data)

}
