% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{vowel}
\alias{vowel}
\title{Create a new vowel object.}
\usage{
vowel(ipa, description)
}
\arguments{
\item{ipa}{The IPA representation of the vowel as a length 1 character
vector. Unicode escape sequences can be utilized for various diacritic marks.}

\item{description}{A description of this vowel's linguistic qualities
as a character vector. Each discrete descriptive quality included must map
to a single element of the vector.

For example, \code{c("close", "front", "unrounded")} would be a valid vector}
}
\value{
A vowel object.
}
\description{
Takes IPA data as a list of character vectors and returns a vowel object.
A vowel object is a subclass of the more general phoneme object.
}
\examples{
a <- vowel("a", c("open", "central", "unrounded"))

}
