% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneme.R
\name{vowel}
\alias{vowel}
\title{Create a new vowel object.}
\usage{
vowel(ipa)
}
\arguments{
\item{ipa}{A named list following the format \code{list(representation = character(), articulation = character())} with each element representing
the vowel's IPA representation (e.g. "a", "p", etc.) and phonetic dimensions respectively.}
}
\value{
A vowel object.
}
\description{
Takes IPA data, formatted as a named list of character vectors, and returns a vowel object. A consonant object is a subclass of the more general phoneme object.
}
\examples{
ipa_data <- list(representation = "e", articulation = c("close-mid", "front", "unrounded"))
e <- vowel(ipa_data)

}
